Add voice notes to every YouTube script-generator path
The brainstorm flow already had a Record Voice Note affordance. Pull the
MediaRecorder + transcription logic into a shared useVoiceNote hook and
VoiceNoteButton component, then wire them into the single-script and
review workflows so every input path supports voice capture.

Also fixes two latent bugs from the original implementation:
- stopRecording now keys off MediaRecorder.state instead of React state,
  so the 120s auto-stop timer no longer no-ops on a stale closure.
- An unmount during recording stops the recorder but a mountedRef guard
  prevents the onstop handler from firing a doomed transcribe request.
