+++ components/shared/index.ts
@@ -23,5 +23,6 @@ export { ThemeProvider } from "./theme-provider"
 export { ThemeToggle } from "./theme-toggle"
 export { UserMenu, type UserMenuUser } from "./user-menu"
 export { UserNotifications } from "./user-notifications"
+export { VoiceNoteButton } from "./voice-note-button"
 export { YouTubeFeedGrid } from "./youtube-feed"
 export { YouTubeFeedSection } from "./youtube-feed-section"
